### 2024-12-11 #7579 @vissarion
> 👋 @vissarion - thanks for your submission. I see a number of issues you should fix before we move on to assigning an editor:
>
>     1. Your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it.
>
>     2. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF.
>
>
> editorialbot commands need to be the first entry in a new comment.

@danielskatz thanks for your quick reaction here! I have fixed all the issues.
