### 2024-12-11 #7579 @olexandr-konovalov
Hello @Viech and @yossibokorbleile - I have found your GitHub usernames in the JOSS reviewers database. Would one or both of you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
