### 2024-12-11 #7579 @Viech
Hi @olexandr-konovalov! I'm afraid my understanding of MCMC methods for integration is too basic to add much to the discussion. (My `convex optimization` keyword is not a good match in that sense; you may want to look for `MCMC`/`monte carlo`, `sampling`, or `(numerical) integration`.)
