### 2024-12-16 #7579 @vissarion
Hi @olexandr-konovalov thank you for taking care of this!

If I may I would like to propose a set of potential reviewers that are not registered in JOSS reviewers database but are authors of highly relevant packages. Please find below a list of github names without tagging:
- yuachen (author of https://github.com/yuachen/polytopewalk)
- qacwnfq (author of https://github.com/modsim/hops)
- YinTat (author of https://github.com/ConstrainedSampler/PolytopeSamplerMatlab)
- pregnault (author of https://github.com/pregnault/samplelim)
