### 2025-02-03 #7579 @pregnault
Hello @olexandr-konovalov,

I am interested in reviewing this submission for JOSS and experimenting on this occasion the original github-based reviewing process. Still, I would need to know what is the deadline for this review process before strictly accepting.
