+++ external/cmake-files/Boost.cmake
@@ -4,7 +4,7 @@ function(GetBoost)
 
 	if (NOT BOOST_DIR)
 
-    set(BOOST_URL "https://boostorg.jfrog.io/artifactory/main/release/1.84.0/source/boost_1_84_0.tar.bz2" CACHE STRING "Boost download URL")
+    set(BOOST_URL "https://archives.boost.io/release/1.84.0/source/boost_1_84_0.tar.bz2" CACHE STRING "Boost download URL")
     set(BOOST_URL_SHA256 "cc4b893acf645c9d4b698e9a0f08ca8846aa5d6c68275c14c3e7949c24109454" CACHE STRING "Boost download URL SHA256 checksum")
 
     include(FetchContent)
