+++ examples/optimization_spectrahedra/Boost.cmake
@@ -2,9 +2,9 @@ function(GetBoost)
 
 	find_path(BOOST_DIR NAMES boost PATHS ../external/)
 
-	if (NOT BOOST_DIR) 
-  
-    set(BOOST_URL "https://boostorg.jfrog.io/artifactory/main/release/1.76.0/source/boost_1_76_0.tar.bz2" CACHE STRING "Boost download URL")
+	if (NOT BOOST_DIR)
+
+    set(BOOST_URL "https://archives.boost.io/release/1.76.0/source/boost_1_76_0.tar.bz2" CACHE STRING "Boost download URL")
     set(BOOST_URL_SHA256 "f0397ba6e982c4450f27bf32a2a83292aba035b827a5623a14636ea583318c41" CACHE STRING "Boost download URL SHA256 checksum")
 
     include(FetchContent)
