+++ .circleci/config.yml
@@ -1,66 +1,0 @@
-# # VolEsti (volume computation and sampling library)
-# # Copyright (c) 2012-2018 Vissarion Fisikopoulos
-# # Licensed under GNU LGPL.3, see LICENCE file
-#
-# version: 2
-# jobs:
-#   build:
-#     docker: # See https://docs.docker.com/get-started/#docker-concepts if you are new to Docker.
-#     #    - image: circleci/ruby:2.4.1-jessie
-#         - image: circleci/ruby:2.6-rc
-#     parallelism: 4
-#     steps:
-#        - checkout
-#        - run:
-#             sudo apt-get update || true;
-#             sudo apt-get install cmake lp-solve;
-#             # sudo apt-get install flex bison libgmp-dev libeigen3-dev coinor-libipopt-dev;
-#             # git clone https://github.com/ethz-adrl/ifopt.git;
-#             # cd ifopt;
-#             # mkdir build;
-#             # cd build;
-#             # cmake ..;
-#             # make;
-#             # sudo make install;
-#             # cd ..;
-#             # cd ..;
-#             # git clone https://github.com/robol/MPSolve;
-#             # cd MPSolve;
-#             # ./autogen.sh;
-#             # ./configure;
-#             # make;
-#             # sudo make install;
-#             # cd ..;
-#             rm -rf build;
-#             mkdir build;
-#             cd build;
-#             cmake ../test -DDISABLE_NLP_ORACLES=ON;
-#             make;
-#             ctest -j8 --verbose;
-
-# VolEsti (volume computation and sampling library)
-# Copyright (c) 2012-2018 Vissarion Fisikopoulos
-# Licensed under GNU LGPL.3, see LICENCE file
-
-version: 2.1
-orbs:
-  codecov: codecov/codecov@3.1.1
-jobs:
-  build:
-    machine:
-      image: ubuntu-2004:202010-01 # recommended linux image
-    resource_class: medium
-    steps:
-       - checkout
-       - run:
-            sudo apt-get update || true;
-            sudo apt-get install cmake lp-solve gcovr;
-            rm -rf build;
-            mkdir build;
-            cd build;
-            cmake -DDISABLE_NLP_ORACLES=ON -DUSE_MKL=OFF -DCODE_COVERAGE=ON ../test;
-            make;
-            ctest -j8 --verbose;
-            gcovr -r ../include . --xml -o coverage.xml;
-       - codecov/upload:
-            file: './build/coverage.xml'
