### 2025-02-10 5007b1d
Sparse optimizations for GaBW sampling (#331)
* simplify gabw sampling

* optimize gabw for sparse polytopes

* minor changes

* add SPARSE boolean for clarity

* compute_reflection_abw_sparse and improve comments
