+++ include/random_walks/gaussian_accelerated_billiard_walk.hpp
@@ -21,6 +21,7 @@
 #include "generators/boost_random_number_generator.hpp"
 #include "sampling/ellipsoid.hpp"
 #include "random_walks/uniform_billiard_walk.hpp"
+#include "random_walks/accelerated_billiard_walk_utils.hpp"
 
 #include "random_walks/compute_diameter.hpp"
 
@@ -47,12 +48,13 @@ struct GaussianAcceleratedBilliardWalk
     struct update_parameters
     {
         update_parameters()
-                :   facet_prev(0), hit_ball(false), inner_vi_ak(0.0), ball_inner_norm(0.0)
+                :   facet_prev(0), hit_ball(false), inner_vi_ak(0.0), ball_inner_norm(0.0), moved_dist(0.0)
         {}
         int facet_prev;
         bool hit_ball;
         double inner_vi_ak;
         double ball_inner_norm;
+        double moved_dist;
     };
 
     parameters param;
@@ -67,9 +69,16 @@ struct GaussianAcceleratedBilliardWalk
     struct Walk
     {
         typedef typename Polytope::PointType Point;
-        typedef typename Polytope::DenseMT DenseMT;
+        typedef typename Polytope::MT MT;
+        typedef typename Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> DenseMT;
         typedef typename Polytope::VT VT;
         typedef typename Point::FT NT;
+        // We do sparse computations iff MT is sparse rowMajor
+        static constexpr bool SPARSE = std::is_same_v<MT, Eigen::SparseMatrix<NT, Eigen::RowMajor>>;
+        // AA is sparse colMajor if MT is sparse rowMajor, and Dense otherwise
+        using AA_type = std::conditional_t< SPARSE, typename Eigen::SparseMatrix<NT>, DenseMT >; 
+        // AE is sparse rowMajor if (MT is sparse rowMajor and E is sparse), and Dense otherwise
+        using AE_type = std::conditional_t< SPARSE && std::is_base_of<Eigen::SparseMatrixBase<E_type>, E_type>::value, typename Eigen::SparseMatrix<NT, Eigen::RowMajor>, DenseMT >; 
 
         void computeLcov(const E_type E)
         {
@@ -110,7 +119,11 @@ struct GaussianAcceleratedBilliardWalk
             _L = compute_diameter<GenericPolytope>::template compute<NT>(P);
             computeLcov(E);
             _E = E;
-            _AA.noalias() = (DenseMT)(P.get_mat() * P.get_mat().transpose());
+            if constexpr (SPARSE) {
+                _AA = (P.get_mat() * P.get_mat().transpose());
+            } else {
+                _AA.noalias() = (P.get_mat() * P.get_mat().transpose());
+            }
             _rho = 1000 * P.dimension(); // upper bound for the number of reflections (experimental)
             initialize(P, p, rng);
         }
@@ -131,7 +144,11 @@ struct GaussianAcceleratedBilliardWalk
                                 ::template compute<NT>(P);
             computeLcov(E);
             _E = E;
-            _AA.noalias() = (DenseMT)(P.get_mat() * P.get_mat().transpose());
+            if constexpr (SPARSE) {
+                _AA = (P.get_mat() * P.get_mat().transpose());
+            } else {
+                _AA.noalias() = (DenseMT)(P.get_mat() * P.get_mat().transpose());
+            }
             _rho = 1000 * P.dimension(); // upper bound for the number of reflections (experimental)
             initialize(P, p, rng);
         }
@@ -162,13 +179,7 @@ struct GaussianAcceleratedBilliardWalk
                 Point p0 = _p;
 
                 it = 0;
-                std::pair<NT, int> pbpair;
[…]

[…]
