+++ README.md
@@ -11,8 +11,6 @@
 ### 🧪 Test results
 
 [![CRAN check](https://badges.cranchecks.info/worst/volesti.svg)](https://cran.r-project.org/web/checks/check_results_volesti.html)
-[![CircleCI master](https://circleci.com/gh/GeomScale/volesti/tree/master.svg?style=shield)](https://circleci.com/gh/GeomScale/volesti/tree/master)
-[![CircleCI develop](https://circleci.com/gh/GeomScale/volesti/tree/develop.svg?style=shield)](https://circleci.com/gh/GeomScale/volesti/tree/develop)
 [![gcc-test](https://github.com/GeomScale/volesti/actions/workflows/cmake-gcc.yml/badge.svg)](https://github.com/GeomScale/volesti/actions/workflows/cmake-gcc.yml?query=branch%3Adevelop)
 [![clang-test](https://github.com/GeomScale/volesti/actions/workflows/cmake-clang.yml/badge.svg)](https://github.com/GeomScale/volesti/actions/workflows/cmake-clang.yml?query=branch%3Adevelop)
 [![codecov](https://codecov.io/gh/GeomScale/volesti/branch/develop/graph/badge.svg)](https://codecov.io/gh/GeomScale/volesti)
