### 2025-02-10 3713bfc
fix macos/ARM/clang compile errors (#297)
* fix cmake bug

* homebrew's lp_solve is not complete. fix LPSolve.cmake

* replace system_clock with high_resolution_clock

* remove header immintrin.h in FloatArray.h

* include <immintrin.h> if the architecture is x64 or x86
