### 2025-02-13 b46eaa0
fix optimization_spectrahedra example
