### 2025-02-13 #7579 @vissarion
Hi @pregnault, according to [JOSS guidelines](https://joss.readthedocs.io/en/latest/sample_messages.html#message-to-reviewers-at-the-start-of-a-review) there is an aim for reviews to be completed within about 2-4 weeks after review process starts (we are currently in pre-review, i.e. searching for reviewers).
