### 2025-02-13 #7579 @olexandr-konovalov
Thank you @pregnault! Indeed, it would be nice to complete like that sample letter says, but if it takes a bit longer, we can be flexible (and I have intensive teaching in the next 4 weeks, so will unlikely come back to this during that time). Please also some information in [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). If this sounds fine, what do you think?
