### 2025-02-14 1cf97e0
fix optimization_spectrahedra example (#344)
* fix optimization_spectrahedra example

* adjust typing
