### 2025-02-17 d7034c6
fixed call_test_max_ball_sparse
