### 2025-02-17 #7579 @olexandr-konovalov
Hello @yuachen and @qacwnfq,

This message is to ask if you can help us out with a submission to JOSS (The Journal of Open Source Software, https://joss.theoj.org), where I’m an editor. Your names were suggested by the authors, since you have contributions to the relevant software, and your expertise may be very helpful.

JOSS publishes articles about open source research software. The submission I'd like you to invite you to review is titled: "volesti: A C++ library for sampling and volume computation on convex bodies", and the submission repository is at https://github.com/GeomScale/volesti (you can find more details in the comments in this issue above).

JOSS is a free, open-source, community driven and developer-friendly online journal (no publisher is seeking to raise revenue from the volunteer labor of researchers!). The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements (see https://joss.readthedocs.io/en/latest/review_criteria.html). Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).

Would one or both of you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
