### 2025-02-20 #7579 @olexandr-konovalov
Thank you for the response and suggestion, @yuachen!

 @bsun1220, your were suggested as potential reviewer of this paper for JOSS - please see several most recent comments in this thread for the details. Would you by any chance be able to do this, please?
