+++ .github/workflows/R-CMD-check-macOS.yml
@@ -20,9 +20,9 @@ jobs:
       fail-fast: false
       matrix:
         config:
-#          - {os: macOS-latest,   r: '4.2.0'} # Error: Package required but not available: ‘RcppEigen’
-#          - {os: macOS-latest,   r: 'devel'} # Error in library(devtools) : there is no package called ‘devtools’
-          - {os: macOS-latest,   r: 'release'}
+          - {os: macos-12, r: 'release'}
+          - {os: macos-13, r: 'release'}
+          - {os: macos-14, r: 'release'}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
