### 2025-02-27 aaeefcd
added CI for macOS
