+++ .github/workflows/cmake-macOS.yml
@@ -1,4 +1,4 @@
-name: cmake-macOS
+name: CI-macOS-AppleClang
 
 on: [push, pull_request]
 
@@ -23,8 +23,15 @@ jobs:
           rm -rf build
           mkdir build
           cd build
+          
+          if [[ "$(uname -s)" == "Darwin" ]]; then
+            CXX_FLAGS="-fsanitize=undefined -g"   # No memory sanitizer on macOS
+          else
+            CXX_FLAGS="-fsanitize=memory -fsanitize=undefined -g"
+          fi
+          
           cmake -D CMAKE_CXX_COMPILER=clang++ \
-                -D CMAKE_CXX_FLAGS="-fsanitize=memory -fsanitize=undefined -g" \
+                -D CMAKE_CXX_FLAGS="${CXX_FLAGS}" \
                 -D DISABLE_NLP_ORACLES=ON \
                 -D USE_MKL=OFF \
                 ../test
