### 2025-02-27 add63b4
Added CI test for macOS
