+++ .github/workflows/cmake-macOS.yml
@@ -17,20 +17,57 @@ jobs:
           - {os: macos-14, compiler: clang++}
           - {os: macos-14, compiler: clang++-16}
           - {os: macos-14, compiler: clang++-17}
-
+    
     runs-on: ${{ matrix.config.os }}
-
+    
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v3
+      
+      # Install Clang versions first
+      - name: Install Required Clang Versions
+        run: |
+          if [[ "${{ matrix.config.compiler }}" == *"clang++-15"* ]]; then
+            brew install llvm@15
+            echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.bash_profile
+            echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.zshrc
+            source $HOME/.bash_profile
+            ln -s /opt/homebrew/opt/llvm@15/bin/clang++ /usr/local/bin/clang++-15 || true
+          fi
+          
+          if [[ "${{ matrix.config.compiler }}" == *"clang++-16"* ]]; then
+            brew install llvm@16
+            echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.bash_profile
+            echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.zshrc
+            source $HOME/.bash_profile
+            ln -s /opt/homebrew/opt/llvm@16/bin/clang++ /usr/local/bin/clang++-16 || true
+          fi
+          
+          if [[ "${{ matrix.config.compiler }}" == *"clang++-17"* ]]; then
+            brew install llvm@17
+            echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.bash_profile
+            echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.zshrc
+            source $HOME/.bash_profile
+            ln -s /opt/homebrew/opt/llvm@17/bin/clang++ /usr/local/bin/clang++-17 || true
+          fi
 
+      # Verify system and compiler after installation
       - name: Verify System & Compiler
         run: |
           echo "System Info:"
           uname -a
           echo "Using Clang version:"
-          ${{ matrix.config.compiler }} --version
+          if [[ "${{ matrix.config.compiler }}" == *"clang++-15"* ]]; then
+            /opt/homebrew/opt/llvm@15/bin/clang++ --version
+          elif [[ "${{ matrix.config.compiler }}" == *"clang++-16"* ]]; then
+            /opt/homebrew/opt/llvm@16/bin/clang++ --version
+          elif [[ "${{ matrix.config.compiler }}" == *"clang++-17"* ]]; then
+            /opt/homebrew/opt/llvm@17/bin/clang++ --version
+          else
+            ${{ matrix.config.compiler }} --version
+          fi
 
+      # Build and test
       - name: Build and Test
         run: |
           rm -rf build
@@ -43,47 +80,50 @@ jobs:
             CXX_FLAGS="-fsanitize=memory -fsanitize=undefined -g"
           fi
           
-          cmake -D CMAKE_CXX_COMPILER=${{ matrix.config.compiler }} \
+          # Use the correct compiler path for LLVM versions
+          COMPILER="${{ matrix.config.compiler }}"
+          if [[ "$COMPILER" == "clang++-15" ]]; then
+            COMPILER="/opt/homebrew/opt/llvm@15/bin/clang++"
+          elif [[ "$COMPILER" == "clang++-16" ]]; then
[…]
