+++ .github/workflows/cmake-macOS.yml
@@ -24,48 +24,44 @@ jobs:
       - name: Checkout Repository
         uses: actions/checkout@v3
       
-      # Install Clang versions first
-      - name: Install Required Clang Versions
+      # Install Clang versions first (for all jobs that need them)
+      - name: Install Clang 15
+        if: contains(matrix.config.compiler, 'clang++-15')
         run: |
-          if [[ "${{ matrix.config.compiler }}" == *"clang++-15"* ]]; then
-            brew install llvm@15
-            echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.bash_profile
-            echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.zshrc
-            source $HOME/.bash_profile
-            ln -s /opt/homebrew/opt/llvm@15/bin/clang++ /usr/local/bin/clang++-15 || true
-          fi
-          
-          if [[ "${{ matrix.config.compiler }}" == *"clang++-16"* ]]; then
-            brew install llvm@16
-            echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.bash_profile
-            echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.zshrc
-            source $HOME/.bash_profile
-            ln -s /opt/homebrew/opt/llvm@16/bin/clang++ /usr/local/bin/clang++-16 || true
-          fi
-          
-          if [[ "${{ matrix.config.compiler }}" == *"clang++-17"* ]]; then
-            brew install llvm@17
-            echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.bash_profile
-            echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.zshrc
-            source $HOME/.bash_profile
-            ln -s /opt/homebrew/opt/llvm@17/bin/clang++ /usr/local/bin/clang++-17 || true
-          fi
+          brew install llvm@15
+          echo "/opt/homebrew/opt/llvm@15/bin" >> $GITHUB_PATH
+          echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.bash_profile
+          echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.zshrc
+          sudo ln -sf /opt/homebrew/opt/llvm@15/bin/clang++ /usr/local/bin/clang++-15
+      
+      - name: Install Clang 16
+        if: contains(matrix.config.compiler, 'clang++-16')
+        run: |
+          brew install llvm@16
+          echo "/opt/homebrew/opt/llvm@16/bin" >> $GITHUB_PATH
+          echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.bash_profile
+          echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.zshrc
+          sudo ln -sf /opt/homebrew/opt/llvm@16/bin/clang++ /usr/local/bin/clang++-16
+      
+      - name: Install Clang 17
+        if: contains(matrix.config.compiler, 'clang++-17') 
+        run: |
+          brew install llvm@17
+          echo "/opt/homebrew/opt/llvm@17/bin" >> $GITHUB_PATH
+          echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.bash_profile
[…]
