+++ .github/workflows/cmake-macOS.yml
@@ -24,48 +24,54 @@ jobs:
       - name: Checkout Repository
         uses: actions/checkout@v3
       
-      # Install Clang versions first (for all jobs that need them)
-      - name: Install Clang 15
-        if: contains(matrix.config.compiler, 'clang++-15')
-        run: |
-          brew install llvm@15
-          echo "/opt/homebrew/opt/llvm@15/bin" >> $GITHUB_PATH
-          echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.bash_profile
-          echo 'export PATH="/opt/homebrew/opt/llvm@15/bin:$PATH"' >> $HOME/.zshrc
-          sudo ln -sf /opt/homebrew/opt/llvm@15/bin/clang++ /usr/local/bin/clang++-15
-      
-      - name: Install Clang 16
-        if: contains(matrix.config.compiler, 'clang++-16')
-        run: |
-          brew install llvm@16
-          echo "/opt/homebrew/opt/llvm@16/bin" >> $GITHUB_PATH
-          echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.bash_profile
-          echo 'export PATH="/opt/homebrew/opt/llvm@16/bin:$PATH"' >> $HOME/.zshrc
-          sudo ln -sf /opt/homebrew/opt/llvm@16/bin/clang++ /usr/local/bin/clang++-16
-      
-      - name: Install Clang 17
-        if: contains(matrix.config.compiler, 'clang++-17') 
-        run: |
-          brew install llvm@17
-          echo "/opt/homebrew/opt/llvm@17/bin" >> $GITHUB_PATH
-          echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.bash_profile
-          echo 'export PATH="/opt/homebrew/opt/llvm@17/bin:$PATH"' >> $HOME/.zshrc
-          sudo ln -sf /opt/homebrew/opt/llvm@17/bin/clang++ /usr/local/bin/clang++-17
-
-      # Verify system and compiler
-      - name: Verify System & Compiler
+      # Prepare LLVM compilers before trying to use them
+      - name: Install LLVM Compilers
         run: |
           echo "System Info:"
           uname -a
-          echo "Using Clang version:"
           
-          # Use the compiler directly without the conditional check
-          # This works because we've installed the compilers and added them to the path
-          ${{ matrix.config.compiler }} --version
-
+          # Install required compiler based on matrix configuration
+          if [[ "${{ matrix.config.compiler }}" == "clang++-15" ]]; then
+            brew install llvm@15
+            # Make sure path exists before verification/build steps
+            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@15/bin/clang++" >> $GITHUB_ENV
+            echo "/opt/homebrew/opt/llvm@15/bin" >> $GITHUB_PATH
+          elif [[ "${{ matrix.config.compiler }}" == "clang++-16" ]]; then
+            brew install llvm@16
+            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@16/bin/clang++" >> $GITHUB_ENV
+            echo "/opt/homebrew/opt/llvm@16/bin" >> $GITHUB_PATH
+          elif [[ "${{ matrix.config.compiler }}" == "clang++-17" ]]; then
+            brew install llvm@17
+            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@17/bin/clang++" >> $GITHUB_ENV
+            echo "/opt/homebrew/opt/llvm@17/bin" >> $GITHUB_PATH
+          else
[…]
