+++ .github/workflows/cmake-macOS.yml
@@ -13,7 +13,7 @@ jobs:
           - {os: macos-12, compiler: clang++-13}
           - {os: macos-12, compiler: clang++-14}
           - {os: macos-13, compiler: clang++}
-          - {os: macos-13, compiler: clang++-15}
+          - {os: macos-13, compiler: clang++-16}
           - {os: macos-14, compiler: clang++}
           - {os: macos-14, compiler: clang++-16}
           - {os: macos-14, compiler: clang++-17}
