+++ .github/workflows/cmake-macOS.yml
@@ -13,7 +13,7 @@ jobs:
           - {os: macos-12, compiler: clang++-13}
           - {os: macos-12, compiler: clang++-14}
           - {os: macos-13, compiler: clang++}
-          - {os: macos-13, compiler: clang++-16}
+          - {os: macos-13, compiler: clang++-15}
           - {os: macos-14, compiler: clang++}
           - {os: macos-14, compiler: clang++-16}
           - {os: macos-14, compiler: clang++-17}
@@ -30,20 +30,41 @@ jobs:
           echo "System Info:"
           uname -a
           
+          # Get architecture info
+          ARCH=$(uname -m)
+          echo "Running on architecture: $ARCH"
+          
           # Install required compiler based on matrix configuration
           if [[ "${{ matrix.config.compiler }}" == "clang++-15" ]]; then
             brew install llvm@15
-            # Make sure path exists before verification/build steps
-            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@15/bin/clang++" >> $GITHUB_ENV
-            echo "/opt/homebrew/opt/llvm@15/bin" >> $GITHUB_PATH
+            if [[ "$ARCH" == "arm64" ]]; then
+              LLVM_PATH=$(brew --prefix llvm@15)
+            else
+              LLVM_PATH="/usr/local/opt/llvm@15"
+            fi
+            echo "COMPILER_PATH=${LLVM_PATH}/bin/clang++" >> $GITHUB_ENV
+            echo "${LLVM_PATH}/bin" >> $GITHUB_PATH
           elif [[ "${{ matrix.config.compiler }}" == "clang++-16" ]]; then
             brew install llvm@16
-            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@16/bin/clang++" >> $GITHUB_ENV
-            echo "/opt/homebrew/opt/llvm@16/bin" >> $GITHUB_PATH
+            if [[ "$ARCH" == "arm64" ]]; then
+              LLVM_PATH=$(brew --prefix llvm@16)
+            else
+              LLVM_PATH="/usr/local/opt/llvm@16"
+            fi
+            echo "COMPILER_PATH=${LLVM_PATH}/bin/clang++" >> $GITHUB_ENV
+            echo "${LLVM_PATH}/bin" >> $GITHUB_PATH
           elif [[ "${{ matrix.config.compiler }}" == "clang++-17" ]]; then
             brew install llvm@17
-            echo "COMPILER_PATH=/opt/homebrew/opt/llvm@17/bin/clang++" >> $GITHUB_ENV
-            echo "/opt/homebrew/opt/llvm@17/bin" >> $GITHUB_PATH
+            if [[ "$ARCH" == "arm64" ]]; then
+              LLVM_PATH=$(brew --prefix llvm@17)
+            else
+              LLVM_PATH="/usr/local/opt/llvm@17"
+            fi
+            echo "COMPILER_PATH=${LLVM_PATH}/bin/clang++" >> $GITHUB_ENV
+            echo "${LLVM_PATH}/bin" >> $GITHUB_PATH
+          elif [[ "${{ matrix.config.compiler }}" == "clang++-13" || "${{ matrix.config.compiler }}" == "clang++-14" ]]; then
+            brew install ${{ matrix.config.compiler }}
+            echo "COMPILER_PATH=${{ matrix.config.compiler }}" >> $GITHUB_ENV
           else
             echo "COMPILER_PATH=${{ matrix.config.compiler }}" >> $GITHUB_ENV
           fi
@@ -52,7 +73,7 @@ jobs:
       - name: Verify Compiler
         run: |
           echo "Using Clang version:"
-          # Use the environment variable set in previous step
[…]
