### 2025-02-28 a4d199d
fixed call_test_max_ball_sparse
