### 2025-03-08 #7886 @editorialbot
[REVIEW]: volesti: A C++ library for sampling and volume computation on convex bodies
**Submitting author:** @vissarion (<a href="http://orcid.org/0000-0002-0780-666X">Vissarion Fisikopoulos</a>)
**Repository:** https://github.com/GeomScale/volesti
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.2.0
