### 2025-03-21 f0bb2eb
sliding window function  minor fix
