### 2025-03-22 f02fb7c
Updated tutorial to reference gen_cube and CRAN v1.1.2
