### 2025-03-26 59f83b1
Fix: Improved burn-in phase handling in sampling functions
