+++ include/sampling/sampling.hpp
@@ -48,13 +48,14 @@ void uniform_sampling(PointList &randPoints,
 
     Point p = starting_point;
 
-    typedef RandomPointGenerator <walk> RandomPointGenerator;
+    typedef RandomPointGenerator <walk> SamplingGenerator;
+    typedef RandomPointGenerator<walk, false> BurnInGenerator;
     if (nburns > 0) {
-        RandomPointGenerator::apply(P, p, nburns, walk_len, randPoints,
+        BurnInGenerator::apply(P, p, nburns, walk_len, randPoints,
                                     push_back_policy, rng);
-        randPoints.clear();
     }
-    RandomPointGenerator::apply(P, p, rnum, walk_len, randPoints,
+
+    SamplingGenerator::apply(P, p, rnum, walk_len, randPoints,
                                 push_back_policy, rng);
 
 
@@ -84,15 +85,16 @@ void uniform_sampling(PointList &randPoints,
 
     //RandomNumberGenerator rng(P.dimension());
     PushBackWalkPolicy push_back_policy;
-    typedef RandomPointGenerator<walk> RandomPointGenerator;
-
+    typedef RandomPointGenerator <walk> SamplingGenerator;
+    typedef RandomPointGenerator<walk, false> BurnInGenerator;
+    
     Point p = starting_point;
     if (nburns > 0) {
-        RandomPointGenerator::apply(P, p, nburns, walk_len, randPoints,
+        BurnInGenerator::apply(P, p, nburns, walk_len, randPoints,
                                     push_back_policy, rng, WalkType.param);
-        randPoints.clear();
+     
     }
-    RandomPointGenerator::apply(P, p, rnum, walk_len, randPoints,
+    SamplingGenerator::apply(P, p, rnum, walk_len, randPoints,
                                 push_back_policy, rng, WalkType.param);
 }
 
@@ -124,14 +126,14 @@ void uniform_sampling_boundary(PointList &randPoints,
 
     Point p = starting_point;
 
-    typedef BoundaryRandomPointGenerator <walk> BoundaryRandomPointGenerator;
+    typedef BoundaryRandomPointGenerator <walk> SamplingGenerator;
+    typedef BoundaryRandomPointGenerator <walk,false> BurnInGenerator;
     if (nburns > 0) {
-        BoundaryRandomPointGenerator::apply(P, p, nburns, walk_len,
+        BurnInGenerator::apply(P, p, nburns, walk_len,
                                             randPoints, push_back_policy, rng);
-        randPoints.clear();
     }
     unsigned int n = rnum / 2;
-    BoundaryRandomPointGenerator::apply(P, p, rnum / 2, walk_len,
+    SamplingGenerator::apply(P, p, rnum / 2, walk_len,
                                         randPoints, push_back_policy, rng);
 
 }
@@ -167,13 +169,13 @@ void gaussian_sampling(PointList &randPoints,
 
     Point p = starting_point;
 
-    typedef GaussianRandomPointGenerator <walk> RandomPointGenerator;
+    typedef GaussianRandomPointGenerator <walk> SamplingGenerator;
+    typedef GaussianRandomPointGenerator <walk,false> BurnInGenerator;
     if (nburns > 0) {
-        RandomPointGenerator::apply(P, p, a, nburns, walk_len, randPoints,
+        BurnInGenerator::apply(P, p, a, nburns, walk_len, randPoints,
                                     push_back_policy, rng);
-        randPoints.clear();
     }
-    RandomPointGenerator::apply(P, p, a, rnum, walk_len, randPoints,
+    SamplingGenerator::apply(P, p, a, rnum, walk_len, randPoints,
[…]
