### 2025-04-02 #7886 @olexandr-konovalov
@bsun1220 thank you for filling in the checklist. Are there any other comments you'd like to add? If you have posted any issues in the submission repository, please cross-reference them here. If you're happy with the paper and recommend acceptance, could you please say it explicitly here.

@pregnault how your review is going - I can see some progress, but a few boxes left to check?
