### 2025-04-02 #7886 @bsun1220
Hey!

Sorry, I thought I left comments but it seems like they did not save. For starters, I really enjoyed using the package. I think the package is highly relevant, performant and also easy to use. Prior to this review, I mainly interacted with Volesti using Dingo (from Python), so I had experience already with the package. All of the downloading / tutorials / tests worked for me and were intuitive. Great job! I am happy with the paper itself and would recommend acceptance.
