### 2025-04-04 5e4b1b7
Fix a typo in cpp_example.md
