### 2025-04-04 #7886 @vissarion
Thanks @pregnault and @bsun1220 for the reviews and feedback.

@pregnault we fixed the typo on the documentation page. Regarding the (theoretical and practical) comparison to other open source packages for polytope sampling (HOPS, PolytopeWalk etc.) that is a project we want to work on. It is very interesting that there are new high quality software packages in this scientific area.
