### 2025-04-04 #7886 @vissarion
@olexandr-konovalov
DOI: 10.5281/zenodo.15146338
release version: v1.2.0
