### 2025-04-05 fabe4a6
Fix #180: Improve efficiency of burn-in phase in sampling functions - Add NoOpWalkPolicy and update tests
