### 2025-04-09 6c800c1
sliding window function minor fix (#351)
* sliding window function  minor fix

* removed whitespaces
