### 2025-04-09 6ec9841
Updated general.md file
