+++ docs/tutorials/general.md
@@ -4,8 +4,8 @@
 
 `volesti` is a `C++` package (with an `R` interface) for computing estimations of volume of polytopes given by a set of points or linear inequalities or Minkowski sum of segments (zonotopes). There are two algorithms for volume estimation and algorithms for sampling, rounding and rotating polytopes.
 
-**Note:** This tutorial is based on package version **1.1.2** from CRAN.  
-Please ensure you have the latest version installed for compatibility.
+**Note:** This tutorial is based on package version **1.2.0** from CRAN.  
+
 
 We can download the `R` package from the [CRAN webpage](https://CRAN.R-project.org/package=volesti).
 
@@ -127,9 +127,9 @@ There are two important parameters `cost per step` and `mixing time` that affect
 library(ggplot2)
 library(volesti)
 for (step in c(1,20,100,150)){
-  for (walk in c("CDHR", "RDHR", "BW")){
+  for (walk in c("CDHR", "RDHR", "BaW")){
     P <- gen_cube(100, 'H')
-    points1 <- sample_points(P, WalkType = walk, walk_step = step, N=1000)
+    points1 = sample_points(P, n = 1000, random_walk = list("walk" = walk, "walk_length" = step))
     g<-plot(ggplot(data.frame( x=points1[1,], y=points1[2,] )) +
 geom_point( aes(x=x, y=y, color=walk)) + coord_fixed(xlim = c(-1,1),
 ylim = c(-1,1)) + ggtitle(sprintf("walk length=%s", step, walk)))
@@ -158,11 +158,11 @@ PV <- gen_cube(3,'V')
 str(PV)
 
 #P = GenRandVpoly(3, 6, body = 'cube')
-tim1 <- system.time({ geom_values = convhulln(PV$V, options = 'FA') })
+tim1 <- system.time({ geom_values = convhulln(PV@V, options = 'FA') })
 tim2 <- system.time({ vol2 = volume(PV) })
 
 cat(sprintf("exact vol = %f\napprx vol = %f\nrelative error = %f\n",
-            geom_values$vol, vol2, abs(geom_values$vol-vol2)/geom_values$vol))
+            geom_values$vol, vol2$volume, abs(geom_values$vol-vol2$volume)/geom_values$vol))
 ```
 
 Now try a higher dimensional example. By setting the `error` parameter we can control the apporximation of the algorithm.
@@ -172,7 +172,7 @@ PH = gen_cube(10,'H')
 volumes <- list()
 for (i in 1:10) {
   # default parameters
-  volumes[[i]] <- volume(PH, error=1)
+  volumes[[i]] <- volume(PH, settings = list("error" = 1))$volume
 }
 options(digits=10)
 summary(as.numeric(volumes))
@@ -181,7 +181,7 @@ summary(as.numeric(volumes))
 ```r
 volumes <- list()
 for (i in 1:10) {
-  volumes[[i]] <- volume(PH, error=0.5)
+  volumes[[i]] <- volume(PH, settings = list("error" = 0.5))$volume
 }
 summary(as.numeric(volumes))
 ```
@@ -191,12 +191,9 @@ Deterministic algorithms for volume are limited to low dimensions (e.g. less tha
 ```r
 library(geometry)
 
-P = GenRandVpoly(15, 30)
-# this will return an error about memory allocation, i.e. the dimension is too high for qhull
[…]
