### 2025-04-11 c8be394
Corrections proposed by JOSS editor
