+++ paper.bib
@@ -218,6 +218,7 @@ keywords = {volume approximation, mathematical software, randomized algorithm, b
   issn = {0387943293  9780387943299  3540943293  9783540943297  038794365X  9780387943657  354094365X  9783540943655},
   publisher = {Springer-Verlag},
   title = {Lectures on polytopes},
+  doi = {10.1007/978-1-4613-8431-1},
   year = 1995
 }
 
@@ -300,7 +301,7 @@ keywords = {Spectahedron, Semidefinite-programming, Sampling, Random walk, Monte
                   {\'{E}}ric Colin de Verdi{\`{e}}re},
   title        = {Geometric Algorithms for Sampling the Flux Space of Metabolic Networks},
   booktitle    = {37th International Symposium on Computational Geometry, SoCG 2021,
-                  June 7-11, 2021, Buffalo, NY, {USA} (Virtual Conference)},
+                  June 7-11, 2021, {Buffalo}, NY, {USA} (Virtual Conference)},
   series       = {LIPIcs},
   volume       = {189},
   pages        = {21:1--21:16},
