### 2025-04-11 #7886 @olexandr-konovalov
@vissarion thanks! I've found two more DOIs for you:

- Lectures on polytopes https://doi.org/10.1007/978-1-4613-8431-1
- Sampling with Riemannian Hamiltonian Monte Carlo i... https://doi.org/10.1074/jbc.R800048200

Could you please update the bibliography?
