### 2025-04-11 d6fc90d
minor changes in paper.bib
