+++ paper.bib
@@ -1,7 +1,7 @@
 @article{Chalkis:2021,
   author = {Apostolos Chalkis and Vissarion Fisikopoulos},
-  title = {{volesti: Volume Approximation and Sampling for Convex
-          Polytopes in R}},
+  title = {{v}olesti: Volume Approximation and Sampling for Convex
+          Polytopes in {R}},
   year = {2021},
   journal = {{The R Journal}},
   doi = {10.32614/RJ-2021-077},
@@ -13,7 +13,7 @@
 
 @article{Chalkis_dingo:2023,
 	author = {Apostolos Chalkis and Vissarion Fisikopoulos and Elias Tsigaridas and Haris Zafeiropoulos},
-	title = {dingo: a Python package for metabolic flux sampling},
+	title = {dingo: a {P}ython package for metabolic flux sampling},
 	elocation-id = {2023.06.18.545486},
 	year = {2023},
 	doi = {10.1101/2023.06.18.545486},
@@ -181,7 +181,7 @@
 
 @article{Chalkis_volume:2023,
 author = {Chalkis, Apostolos and Emiris, Ioannis Z. and Fisikopoulos, Vissarion},
-title = {A Practical Algorithm&nbsp;for Volume Estimation based on Billiard Trajectories and Simulated Annealing},
+title = {A Practical Algorithm for Volume Estimation based on Billiard Trajectories and Simulated Annealing},
 year = {2023},
 issue_date = {December 2023},
 publisher = {Association for Computing Machinery},
@@ -252,7 +252,7 @@ keywords = {Hybrid probabilistic inference, Weighted Model Integration, Satisfia
  editor = {S. Koyejo and S. Mohamed and A. Agarwal and D. Belgrave and K. Cho and A. Oh},
  pages = {31684--31696},
  publisher = {Curran Associates, Inc.},
- title = {Sampling with Riemannian Hamiltonian Monte Carlo in a Constrained Space},
+ title = {Sampling with {R}iemannian {H}amiltonian {M}onte {C}arlo in a Constrained Space},
  url = {https://proceedings.neurips.cc/paper_files/paper/2022/file/cdaa7f07b0c5a7803927d20aa717132e-Paper-Conference.pdf},
  volume = {35},
  year = {2022}
@@ -260,7 +260,7 @@ keywords = {Hybrid probabilistic inference, Weighted Model Integration, Satisfia
 
 @article{Chalkis_hmc:2023,
 author = {Chalkis, Apostolos and Fisikopoulos, Vissarion and Papachristou, Marios and Tsigaridas, Elias},
-title = {Truncated Log-concave Sampling for Convex Bodies with Reflective Hamiltonian Monte Carlo},
+title = {Truncated Log-concave Sampling for Convex Bodies with Reflective {H}amiltonian {M}onte {C}arlo},
 year = {2023},
 issue_date = {June 2023},
 publisher = {Association for Computing Machinery},
