### 2025-04-11 6750dc1
minor changes for JOSS publication (#356)
* minor changes in paper.md

* minor changes in paper.bib
