### 2025-04-11 ca6721b
Add countries in JOSS affiliations
