### 2025-04-11 #7886 @danielskatz
Congratulations to @vissarion (Vissarion Fisikopoulos) and co-authors on your JOSS publication!!

JOSS depends on volunteers and couldn't do this without you
